Fix: Use BigInteger for Telegram user IDs

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, Float, DateTime, ForeignKey
+from sqlalchemy import create_engine, Column, Integer, String, Float, DateTime, ForeignKey, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from datetime import datetime
@@ -9,10 +9,10 @@ class User(Base):
     __tablename__ = 'users'
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)
     username = Column(String)
     balance = Column(Float, default=10000.0)  # Initial balance 10k USD
-    referred_by = Column(Integer, nullable=True)
+    referred_by = Column(BigInteger, nullable=True)
     created_at = Column(DateTime, default=datetime.utcnow)
     
     trades = relationship("Trade", back_populates="user", cascade="all, delete-orphan")
